+++ R/global_imports.R
@@ -3,7 +3,7 @@
 #' @importFrom magrittr %>%
 #' @importFrom dplyr .data
 #' @importFrom loo loo waic loo_compare
-#' @importFrom stats gaussian binomial mad median
+#' @importFrom stats gaussian binomial median
 #' @importFrom rlang !! := .env
 "_PACKAGE"